+++ Seaurchin/ScenePlayer.cpp
@@ -93,7 +93,7 @@ void ScenePlayer::LoadWorker()
     analyzer->LoadFromFile(scorefile);
     analyzer->RenderScoreData(data);
     for (auto &note : data) {
-        if (note->Type.test(SusNoteType::Slide) || note->Type.test(SusNoteType::AirAction)) CalculateCurves(note);
+        if (note->Type.test((size_t)SusNoteType::Slide) || note->Type.test((size_t)SusNoteType::AirAction)) CalculateCurves(note);
     }
     processor->Reset();
     State = PlayingState::BgmNotLoaded;
@@ -121,8 +121,8 @@ void ScenePlayer::CalculateCurves(std::shared_ptr<SusDrawableNoteData> note)
 
     controlPoints.push_back(make_tuple(0, (lastStep->StartLane + lastStep->Length / 2.0) / 16.0));
     for (auto &slideElement : note->ExtraData) {
-        if (slideElement->Type.test(SusNoteType::ExTap)) continue;
-        if (slideElement->Type.test(SusNoteType::Control)) {
+        if (slideElement->Type.test((size_t)SusNoteType::ExTap)) continue;
+        if (slideElement->Type.test((size_t)SusNoteType::Control)) {
             auto cpi = make_tuple(slideElement->StartTime - lastStep->StartTime, (slideElement->StartLane + slideElement->Length / 2.0) / 16.0);
             controlPoints.push_back(cpi);
             continue;
